Move api/server/middleware to daemon/server/middleware
