+++ api/server/middleware.go
@@ -3,7 +3,7 @@ package server
 import (
 	"github.com/containerd/log"
 	"github.com/docker/docker/api/server/httputils"
-	"github.com/docker/docker/api/server/middleware"
+	"github.com/docker/docker/daemon/server/middleware"
 )
 
 // handlerWithGlobalMiddlewares wraps the handler function for a request with
