+++ daemon/command/daemon.go
@@ -21,7 +21,6 @@ import (
 	"github.com/docker/docker/api"
 	apiserver "github.com/docker/docker/api/server"
 	buildbackend "github.com/docker/docker/api/server/backend/build"
-	"github.com/docker/docker/api/server/middleware"
 	buildkit "github.com/docker/docker/builder/builder-next"
 	"github.com/docker/docker/builder/builder-next/exporter"
 	"github.com/docker/docker/builder/dockerfile"
@@ -31,6 +30,7 @@ import (
 	"github.com/docker/docker/daemon/command/trap"
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/listeners"
+	"github.com/docker/docker/daemon/server/middleware"
 	"github.com/docker/docker/daemon/server/router"
 	"github.com/docker/docker/daemon/server/router/build"
 	checkpointrouter "github.com/docker/docker/daemon/server/router/checkpoint"
