+++ daemon/server/middleware/middleware.go
