+++ daemon/server/middleware/version.go
