+++ daemon/server/middleware/version_test.go
